Refactor to detect sflow enabled interfaces
