+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/intended/configs/sflow-tests-l2-leaf2.cfg
@@ -23,6 +23,7 @@ sflow destination 10.10.10.15
 sflow destination 10.10.10.16
 sflow destination 10.10.10.17
 sflow source-interface Vlan4092
+sflow run
 !
 no enable password
 no aaa root
