+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/intended/configs/sflow-tests-leaf1.cfg
@@ -19,6 +19,7 @@ sflow vrf sflowvrf destination 10.10.10.17
 sflow vrf sflowvrf source-interface Loopback321
 sflow destination 10.10.10.10
 sflow destination 10.10.10.11 1234
+sflow run
 !
 spanning-tree mode mstp
 spanning-tree mst 0 priority 4096
