+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/intended/configs/sflow-tests-leaf4.cfg
@@ -19,6 +19,7 @@ sflow vrf sflowvrf destination 10.10.10.17
 sflow vrf sflowvrf source-interface Loopback321
 sflow destination 10.10.10.10
 sflow destination 10.10.10.11 1234
+sflow run
 !
 spanning-tree mode mstp
 no spanning-tree vlan-id 4093-4094
