+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/intended/structured_configs/sflow-tests-l2-leaf1.yml
@@ -27,29 +27,6 @@ management_api_http:
   enable_vrfs:
   - name: MGMT
   enable_https: true
-sflow:
-  destinations:
-  - destination: 10.10.10.10
-  - destination: 10.10.10.11
-    port: 1234
-  - destination: 10.10.10.15
-  - destination: 10.10.10.16
-  - destination: 10.10.10.17
-  source_interface: Vlan4092
-  vrfs:
-  - name: sflowvrf
-    destinations:
-    - destination: 10.10.10.12
-    - destination: 10.10.10.14
-    - destination: 10.10.10.16
-    - destination: 10.10.10.17
-    source_interface: Loopback321
-  - name: MGMT
-    destinations:
-    - destination: 10.10.10.13
-    - destination: 10.10.10.14
-    - destination: 10.10.10.17
-    source_interface: Management1
 ethernet_interfaces:
 - name: Ethernet1
   peer: sflow-tests-leaf3
@@ -96,5 +73,29 @@ vlan_interfaces:
   mtu: 1500
   ip_address: 10.254.254.4/24
   type: inband_mgmt
+sflow:
+  run: true
+  destinations:
+  - destination: 10.10.10.10
+  - destination: 10.10.10.11
+    port: 1234
+  - destination: 10.10.10.15
+  - destination: 10.10.10.16
+  - destination: 10.10.10.17
+  source_interface: Vlan4092
+  vrfs:
+  - name: sflowvrf
+    destinations:
+    - destination: 10.10.10.12
+    - destination: 10.10.10.14
+    - destination: 10.10.10.16
+    - destination: 10.10.10.17
+    source_interface: Loopback321
+  - name: MGMT
+    destinations:
+    - destination: 10.10.10.13
+    - destination: 10.10.10.14
+    - destination: 10.10.10.17
+    source_interface: Management1
 metadata:
   platform: vEOS-lab
