+++ ansible_collections/arista/avd/roles/eos_designs/python_modules/sflow/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2023 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
+from .avdstructuredconfig import AvdStructuredConfigSflow
+
+__all__ = ["AvdStructuredConfigSflow"]
