+++ ansible_collections/arista/avd/roles/eos_designs/schemas/schema_fragments/fabric_sflow.schema.yml
@@ -1,3 +1,6 @@
+# Copyright (c) 2023 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
 # yaml-language-server: $schema=../../../../plugins/plugin_utils/schema/avd_meta_schema.json
 # Line above is used by RedHat's YAML Schema vscode extension
 # Use Ctrl + Space to get suggestions for every field. Autocomplete will pop up after typing 2 letters.
@@ -10,6 +13,7 @@ keys:
     description: |-
       sFlow settings.
       Configure destinations and default enabling of sFlow for varous interface types across the fabric.
+      The sFlow process will be enabled if any interface is enabled for sFlow.
       For source-interfaces see "source_interfaces.sflow"
     keys:
       destinations:
